Refactor package name for emoji-updater and fix javadoc
