+++ emoji-updater/pom.xml
@@ -12,7 +12,7 @@
         <maven.compiler.release>11</maven.compiler.release>
         <connect.version>2.0.1</connect.version>
         <javafx.maven.plugin.version>0.0.8</javafx.maven.plugin.version>
-        <main.class>com.gluonhq.emoji.tools.Main</main.class>
+        <main.class>com.gluonhq.emoji.updater.Main</main.class>
     </properties>
 
     <dependencies>
