+++ emoji-updater/src/main/java/com/gluonhq/emoji/updater/Emoji.java
@@ -1,4 +1,4 @@
-package com.gluonhq.emoji.tools;
+package com.gluonhq.emoji.updater;
 
 import java.util.List;
